third_party/blink/web_tests/platform/generic/fast/dom/HTMLScriptElement/script-sync-load-failure-expected.txt

CONSOLE MESSAGE: success.js loaded
Sync scripts that fail to load should not block later sync scripts from executing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
